requirements.txt
Django==5.0.7
djangorestframework==3.15.2
pytest==8.1.1
pytest==8.3.2
